Add files via upload
A smart contract written in Solidity with the help of which a manager can create auctions in which bidders can bid for a property. The manager conducts and concludes the auction after which the money goes to the owner of the property listed in the auction. The manager also decides the minimum highest bidding bid.  The bidders need to pay a minimum amount of fee to enter the auction after which they can compete for the property. The highest bidder has money deducted from their account and the remaining bidders get their money back. The highest bidder gets back the difference between the highest bid they placed and the highest binding bid.